fix : 납부일 자동 알림 로직 수정

+++ src/main/java/com/sosim/server/event/domain/entity/Event.java
@@ -106,7 +106,7 @@ public class Event extends BaseTimeEntity {
     }
 
     public boolean included(Group group) {
-        return group.getId() == this.group.getId();
+        return group.getId().longValue() == this.group.getId().longValue();
     }
 
     private void validSituation(Situation situation) {
